requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pygame>=2.5.0
gTTS>=2.3.2
PyAudio>=0.2.13
